+++ engine/updater/Manifest.php
@@ -1,0 +1,12 @@
+<?php
+namespace Updater;
+
+class Manifest
+{
+    public function __construct(public array $data) {}
+
+    public function channel(string $name): array
+    {
+        return (array)($this->data['channels'][$name] ?? []);
+    }
+}